Separate apollo client out again
